+++ packages/at_commons/test/enroll_params_test.dart
@@ -1,5 +1,5 @@
 import 'dart:convert';
-import 'package:at_commons/src/verb/enroll_params.dart';
+import 'package:at_commons/at_commons.dart';
 import 'package:test/test.dart';
 
 void main() {
@@ -52,6 +52,21 @@ void main() {
       var enrollParams = jsonDecode(command);
       expect(enrollParams['enrollmentId'], '123');
     });
+
+    test('A test to verify enroll list regex with params', () {
+      String command =
+          'enroll:list:{"enrollmentStatusFilter":"[pending, approved]"}';
+      expect(RegExp(VerbSyntax.enroll).hasMatch(command), true);
+
+      command = command.replaceAll('enroll:list:', '');
+      var enrollParams = jsonDecode(command);
+      expect(enrollParams['enrollmentStatusFilter'], '[pending, approved]');
+    });
+
+    test('A test to verify enroll list regex without params', () {
+      String command = 'enroll:list';
+      expect(RegExp(VerbSyntax.enroll).hasMatch(command), true);
+    });
   });
 
   group('A group of tests to verify toJson and fromJson in EnrollParams', () {
@@ -64,7 +79,11 @@ void main() {
         ..enrollmentId = '1234'
         ..encryptedAPKAMSymmetricKey = 'dummy_pkam_sym_key'
         ..encryptedDefaultEncryptionPrivateKey = 'dummy_encrypted_private_key'
-        ..encryptedDefaultSelfEncryptionKey = 'dummy_self_encryption_key';
+        ..encryptedDefaultSelfEncryptionKey = 'dummy_self_encryption_key'
+        ..enrollmentStatusFilter = [
+          EnrollmentStatus.approved,
+          EnrollmentStatus.pending
+        ];
 
       Map<String, dynamic> enrollParamsMap = enrollParams.toJson();
       expect(enrollParamsMap['appName'], 'wavi');
@@ -78,6 +97,8 @@ void main() {
           'dummy_encrypted_private_key');
       expect(enrollParamsMap['encryptedDefaultSelfEncryptionKey'],
           'dummy_self_encryption_key');
+      expect(
+          enrollParamsMap['enrollmentStatusFilter'], ['approved', 'pending']);
     });
 
     test('A test to verify fromJson', () {
@@ -93,6 +114,7 @@ void main() {
       enrollParamsMap['encryptedDefaultSelfEncryptionKey'] =
           'dummy_self_encryption_key';
       enrollParamsMap['otp'] = '123';
+      enrollParamsMap['enrollmentStatusFilter'] = ['approved', 'pending'];
 
       var enrollParams = EnrollParams.fromJson(enrollParamsMap);
       expect(enrollParams.appName, 'wavi');
@@ -106,6 +128,8 @@ void main() {
           'dummy_self_encryption_key');
       expect(enrollParams.otp, '123');
       expect(enrollParams.namespaces, {'wavi': 'rw', '__manage': 'r'});
+      expect(enrollParams.enrollmentStatusFilter,
+          [EnrollmentStatus.approved, EnrollmentStatus.pending]);
     });
   });
 }