+++ packages/at_commons/CHANGELOG.md
@@ -1,6 +1,5 @@
 ## 4.0.4
-- feat: update enroll regex so that enroll:list accepts optional params
-- feat: introduce field in EnrollParams named 'enrollmentStatusFilter'
+- feat: Enhance enroll:list to enable filtering of enrollments based on their status
 ## 4.0.3
 - fix: "toJson()" invoked on "pubKeyHash" leads to NullPointerException.
 ## 4.0.2