+++ packages/at_commons/CHANGELOG.md
@@ -1,3 +1,5 @@
+## 4.0.4
+- feat: Enhance enroll:list to enable filtering based on enrollment status
 ## 4.0.3
 - fix: "toJson()" invoked on "pubKeyHash" leads to NullPointerException.
 ## 4.0.2